Organizing code to be shared across llm tasks
