+++ llm/inference/transformer.py
@@ -1,32 +1,14 @@
 from __future__ import annotations
-from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
 import gc
 
-from queue import Queue
-from multiprocessing import Pipe, Process, set_start_method
-from threading import Thread
-from typing import Any, Callable, Iterable, List, Optional, Tuple, Union
+from typing import Iterable, List, Optional, Tuple, Union
 
 import torch
 from transformers import PreTrainedModel, PreTrainedTokenizerBase, LogitsProcessorList, LogitsProcessor, LogitsWarper, RepetitionPenaltyLogitsProcessor, TemperatureLogitsWarper, TopPLogitsWarper, TopKLogitsWarper
 
-from llm.qa.model_configs import ModelConfig
-
-
-class InferenceEngine(ABC):
-    @abstractmethod
-    def generate_stream(self, prompt: str, **kwargs) -> Iterable[str]:
-        """
-        Stream generated text as each new token is added
-        """
-        raise NotImplementedError()
-
-    def get_answer(self, prompt: str, **kwargs) -> str:
-        answer = ""
-        for _answer in self.generate_stream(prompt, **kwargs):
-            answer = _answer
-        return answer
+from llm.inference.base import InferenceEngine
+from llm.model_configs import ModelConfig
 
 
 class TransformersEngine(InferenceEngine):
@@ -58,11 +40,12 @@ class TransformersEngine(InferenceEngine):
         stop_token_ids: Optional[List[int]] = None,
         stop_str: Union[str, List[str]] = "",
         logits_config: Optional[LogitsProcessorConfig] = None,
+        **kwargs,
     ) -> Iterable[str]:
         logits_processor: Optional[LogitsProcessorList] = None
         if logits_config:
             logits_processor = logits_config.load()
-            do_sampling = logits_config.do_sampling
+            do_sampling = True if logits_config.do_sampling else False
         else:
             do_sampling = True
 
@@ -150,7 +133,7 @@ class TransformersEngine(InferenceEngine):
         In subsequent steps, attention is only calculated for new tokens.
         """
         input_tensor = torch.as_tensor([input_ids], device=self.model.device)
-        encoder_output: Optional[torch.Tensor] = None
+        encoder_output: Optional[Tuple[torch.FloatTensor]] = None
         if self.model.config.is_encoder_decoder:
             # Encoder-Decoder models
             encoder_output = self.model.encoder(input_ids=input_tensor)[0]
@@ -234,127 +217,6 @@ class TransformersEngine(InferenceEngine):
         # Avoid decoding tokens until the final step
         kwargs.setdefault("stream_interval", -1)
         return super().get_answer(prompt, **kwargs)
-
-
-class MultiprocessEngine(InferenceEngine):
-    """
-    Run inference engines in background processes with queued requests/responses
-
-    Enables thread-safe non-blocking inference across multiple devices
-    """
-    def __init__(self, workers: List[WorkerPipe]):
-        self.workers = workers
-        self.closed = False
-        self.queue: Queue[WorkerPipe] = Queue(len(self.workers))
-        for worker in self.workers:
-            self.queue.put(worker)
-
-    @classmethod
-    def from_model_config(cls, model_config: ModelConfig, num_workers: Optional[int] = None, wait: bool = True) -> MultiprocessEngine:
-        # Required for CUDA
-        set_start_method("spawn")
-
-        if num_workers is None:
-            # TODO: Should we also check CUDA_VISIBLE_DEVICES?
-            num_workers = torch.cuda.device_count()
-
-        workers: List[WorkerPipe] = []
-        for i in range(num_workers):
-            worker = WorkerPipe()
-            watch_proc = Process(
-                None,
-                cls._transformers_worker,
-                args=[worker, model_config, i],
-                kwargs={"signal_ready": wait},
-                daemon=True,
-            )
-            watch_proc.start()
-            workers.append(worker)
-
-        if wait:
-            for worker in workers:
-                # Collect initial sentinel values indicating the workers have loaded
-                worker.get_response()
-
-        return cls(workers)
-
-    @staticmethod
-    def _transformers_worker(pipe: WorkerPipe, model_config: ModelConfig, local_rank: int, signal_ready: bool = False):
-        engine = TransformersEngine.from_model_config(model_config, device_map={"": local_rank})
-        if signal_ready:
-            pipe.send_response(None)
-        while True:
-            request = pipe.get_request()
-            for text in engine.generate_stream(request.prompt, **request.kwargs):
-                pipe.send_response(text)
-
-            # None indicates to the requester that the stream is completed
-            pipe.send_response(None)
-
-    def generate_stream(
-        self,
-        prompt: str,
-        **kwargs,
-    ) -> Iterable[str]:
-        request = StreamRequest(prompt, **kwargs)
-        # Wait for a worker to be available
-        worker = self.queue.get()
-        try:
-            worker.send_request(request)
-            while True:
-                # Yield results until sentinel value is recieved
-                text = worker.get_response()
-                if text is not None:
-                    yield text
-                else:
-                    # Stream completed
-                    break
-
-            self.queue.task_done()
-        finally:
-            if not self.closed:
-                # Add pipe back to queue to process new requests
-                self.queue.put(worker)
-
-    def close(self):
-        self.closed = True
-        for pipe in self.workers:
-            pipe.close()
-
-
-class StreamRequest:
-    """
-    Wraps an inference request to be processed by a worker
-    """
-    def __init__(self, prompt: str, **kwargs) -> None:
-        self.prompt = prompt
-        self.kwargs = kwargs
-
-
-class WorkerPipe:
-    """
-    Manages communication between an inference worker and the main process
-    """
-    def __init__(self):
-        self.parent_conn, self.child_conn = Pipe()
-
-    def close(self):
-        self.parent_conn.close()
-        self.child_conn.close()
-
-    ### Main proc methods ####
-    def send_request(self, request: StreamRequest):
-        self.parent_conn.send(request)
-
-    def get_response(self) -> Optional[str]:
-        return self.parent_conn.recv()
-
-    ### Worker methods ####
-    def get_request(self) -> StreamRequest:
-        return self.child_conn.recv()
-
-    def send_response(self, response: Optional[str]):
-        self.child_conn.send(response)
 
 
 @dataclass
